requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.0.1
django-cors-headers==3.3.0
django-environ==0.4.5
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
